backend/server/requirements.txt
Django==3.2.13
environs==7.3.1
django-cors-headers
psycopg2-binary
djangorestframework
jwt